Revert "renaming to lower case file in Action, Backend and Environment"
This reverts commit dadf84d4b3bef34dcfdf651c606fe3d1d643e7f6.
