+++ grid2op/Action/PowerlineSetAndDispatchAction.py
@@ -6,22 +6,31 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
-from grid2op.Action.playableAction import PlayableAction
+from grid2op.Action.PlayableAction import PlayableAction
 
 
-class TopologyChangeAction(PlayableAction):
+class PowerlineSetAndDispatchAction(PlayableAction):
     """
-    This type of :class:`PlayableAction` implements the modifications
-    of the grid with "change" topological actions.
+    This type of :class:`PlayableAction` only implements the
+    modifications of the grid with set topological and dispatch actions.
 
-    It accepts the key words: "change_line_status" and "change_bus".
+    It accepts the key words: "set_line_status" and  "redispatch".
     Nothing else is supported and any attempt to use something else
     will have no impact.
     """
 
-    authorized_keys = {"change_line_status", "change_bus"}
+    authorized_keys = {
+        "set_line_status",
+        # "set_bus",
+        "redispatch",
+    }
 
-    attr_list_vect = ["_change_bus_vect", "_switch_line_status"]
+    attr_list_vect = [
+        "_set_line_status",
+        # "_set_topo_vect",
+        "_redispatch",
+    ]
+
     attr_list_set = set(attr_list_vect)
 
     def __init__(self):
