+++ grid2op/Action/TopologyAction.py
@@ -6,26 +6,32 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
-from grid2op.Action.playableAction import PlayableAction
+from grid2op.Action.PlayableAction import PlayableAction
 
 
-class TopologyChangeAndDispatchAction(PlayableAction):
+class TopologyAction(PlayableAction):
     """
-    This type of :class:`PlayableAction` implements the modifications
-    of the grid with "change" topological actions and allows for redispatching.
+    This type of :class:`PlayableAction` only implements the
+    modifications of the grid with topological actions.
 
-    It accepts the key words: "change_line_status", "change_bus"
-    and "redispatch".
+    It accepts the key words: "set_line_status", "change_line_status",
+    "set_bus" and "change_bus".
     Nothing else is supported and any attempt to use something else
     will have no impact.
     """
 
-    authorized_keys = {"change_line_status", "change_bus", "redispatch"}
+    authorized_keys = {
+        "set_line_status",
+        "change_line_status",
+        "set_bus",
+        "change_bus",
+    }
 
     attr_list_vect = [
-        "_change_bus_vect",
+        "_set_line_status",
         "_switch_line_status",
-        "_redispatch",
+        "_set_topo_vect",
+        "_change_bus_vect",
     ]
 
     attr_list_set = set(attr_list_vect)
