+++ grid2op/Action/TopologySetAction.py
@@ -6,22 +6,23 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
-from grid2op.Action.playableAction import PlayableAction
+from grid2op.Action.PlayableAction import PlayableAction
 
 
-class PowerlineChangeAndDispatchAction(PlayableAction):
+class TopologySetAction(PlayableAction):
     """
-    This type of :class:`PlayableAction` only implements the
-    modifications of the grid with powerlines switch and dispatch actions.
+    This type of :class:`PlayableAction` implements the modifications
+    of the grid with "set" topological actions.
 
-    It accepts the key words: "change_line_status" and "redispatch".
+    It accepts the key words: "set_line_status" and "set_bus".
     Nothing else is supported and any attempt to use something else
     will have no impact.
     """
 
-    authorized_keys = {"change_line_status", "redispatch"}
+    authorized_keys = {"set_line_status", "set_bus"}
 
-    attr_list_vect = ["_switch_line_status", "_redispatch"]
+    attr_list_vect = ["_set_line_status", "_set_topo_vect"]
+
     attr_list_set = set(attr_list_vect)
 
     def __init__(self):
