+++ grid2op/Action/TopologySetAndDispatchAction.py
@@ -6,31 +6,22 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
-from grid2op.Action.playableAction import PlayableAction
+from grid2op.Action.PlayableAction import PlayableAction
 
 
-class PowerlineSetAndDispatchAction(PlayableAction):
+class TopologySetAndDispatchAction(PlayableAction):
     """
-    This type of :class:`PlayableAction` only implements the
-    modifications of the grid with set topological and dispatch actions.
+    This type of :class:`PlayableAction` implements the modifications
+    of the grid with "set" topological actions and allows for redispatching.
 
-    It accepts the key words: "set_line_status" and  "redispatch".
+    It accepts the key words: "set_line_status", "set_bus" and "redispatch".
     Nothing else is supported and any attempt to use something else
     will have no impact.
     """
 
-    authorized_keys = {
-        "set_line_status",
-        # "set_bus",
-        "redispatch",
-    }
+    authorized_keys = {"set_line_status", "set_bus", "redispatch"}
 
-    attr_list_vect = [
-        "_set_line_status",
-        # "_set_topo_vect",
-        "_redispatch",
-    ]
-
+    attr_list_vect = ["_set_line_status", "_set_topo_vect", "_redispatch"]
     attr_list_set = set(attr_list_vect)
 
     def __init__(self):
