+++ grid2op/Action/__init__.py
@@ -23,29 +23,29 @@ __all__ = [
 ]
 
 # Internals
-from grid2op.Action.baseAction import BaseAction
-from grid2op.Action.playableAction import PlayableAction
-from grid2op.Action.voltageOnlyAction import VoltageOnlyAction
-from grid2op.Action.completeAction import CompleteAction
-from grid2op.Action.actionSpace import ActionSpace
-from grid2op.Action.serializableActionSpace import SerializableActionSpace
+from grid2op.Action.BaseAction import BaseAction
+from grid2op.Action.PlayableAction import PlayableAction
+from grid2op.Action.VoltageOnlyAction import VoltageOnlyAction
+from grid2op.Action.CompleteAction import CompleteAction
+from grid2op.Action.ActionSpace import ActionSpace
+from grid2op.Action.SerializableActionSpace import SerializableActionSpace
 
-from grid2op.Action.dontAct import DontAct
-from grid2op.Action.powerlineSetAction import PowerlineSetAction
-from grid2op.Action.powerlineChangeAction import PowerlineChangeAction
-from grid2op.Action.powerlineSetAndDispatchAction import PowerlineSetAndDispatchAction
-from grid2op.Action.powerlineChangeAndDispatchAction import (
+from grid2op.Action.DontAct import DontAct
+from grid2op.Action.PowerlineSetAction import PowerlineSetAction
+from grid2op.Action.PowerlineChangeAction import PowerlineChangeAction
+from grid2op.Action.PowerlineSetAndDispatchAction import PowerlineSetAndDispatchAction
+from grid2op.Action.PowerlineChangeAndDispatchAction import (
     PowerlineChangeAndDispatchAction,
 )
-from grid2op.Action.powerlineChangeDispatchAndStorageAction import (
+from grid2op.Action.PowerlineChangeDispatchAndStorageAction import (
     PowerlineChangeDispatchAndStorageAction,
 )
-from grid2op.Action.topologyAction import TopologyAction
-from grid2op.Action.topologyAndDispatchAction import TopologyAndDispatchAction
-from grid2op.Action.topologySetAction import TopologySetAction
-from grid2op.Action.topologySetAndDispatchAction import TopologySetAndDispatchAction
-from grid2op.Action.topologyChangeAction import TopologyChangeAction
-from grid2op.Action.topologyChangeAndDispatchAction import (
+from grid2op.Action.TopologyAction import TopologyAction
+from grid2op.Action.TopologyAndDispatchAction import TopologyAndDispatchAction
+from grid2op.Action.TopologySetAction import TopologySetAction
+from grid2op.Action.TopologySetAndDispatchAction import TopologySetAndDispatchAction
+from grid2op.Action.TopologyChangeAction import TopologyChangeAction
+from grid2op.Action.TopologyChangeAndDispatchAction import (
     TopologyChangeAndDispatchAction,
 )
-from grid2op.Action.dispatchAction import DispatchAction
+from grid2op.Action.DispatchAction import DispatchAction
