+++ grid2op/Backend/PandaPowerBackend.py
@@ -18,7 +18,7 @@ import pandapower as pp
 import scipy
 
 from grid2op.dtypes import dt_int, dt_float, dt_bool
-from grid2op.Backend.backend import Backend
+from grid2op.Backend.Backend import Backend
 from grid2op.Action import BaseAction
 from grid2op.Exceptions import *
