+++ grid2op/Backend/__init__.py
@@ -1,4 +1,4 @@
 __all__ = ["Backend", "PandaPowerBackend"]
 
-from grid2op.Backend.backend import Backend
-from grid2op.Backend.pandaPowerBackend import PandaPowerBackend
+from grid2op.Backend.Backend import Backend
+from grid2op.Backend.PandaPowerBackend import PandaPowerBackend
