+++ grid2op/Environment/MultiEnvMultiProcess.py
@@ -11,7 +11,7 @@ import numpy as np
 from grid2op.dtypes import dt_int
 from grid2op.Exceptions import Grid2OpException, MultiEnvException
 from grid2op.Space import GridObjects
-from grid2op.Environment.baseMultiProcessEnv import BaseMultiProcessEnvironment
+from grid2op.Environment.BaseMultiProcessEnv import BaseMultiProcessEnvironment
 from grid2op.Action import BaseAction
