+++ grid2op/Environment/__init__.py
@@ -8,10 +8,10 @@ __all__ = [
     "TimedOutEnvironment"
 ]
 
-from grid2op.Environment.baseEnv import BaseEnv
-from grid2op.Environment.environment import Environment
-from grid2op.Environment.baseMultiProcessEnv import BaseMultiProcessEnvironment
-from grid2op.Environment.singleEnvMultiProcess import SingleEnvMultiProcess
-from grid2op.Environment.multiEnvMultiProcess import MultiEnvMultiProcess
-from grid2op.Environment.multiMixEnv import MultiMixEnvironment
+from grid2op.Environment.BaseEnv import BaseEnv
+from grid2op.Environment.Environment import Environment
+from grid2op.Environment.BaseMultiProcessEnv import BaseMultiProcessEnvironment
+from grid2op.Environment.SingleEnvMultiProcess import SingleEnvMultiProcess
+from grid2op.Environment.MultiEnvMultiProcess import MultiEnvMultiProcess
+from grid2op.Environment.MultiMixEnv import MultiMixEnvironment
 from grid2op.Environment.timedOutEnv import TimedOutEnvironment
