+++ grid2op/Reward/_alarmScore.py
@@ -65,7 +65,14 @@ class _AlarmScore(AlarmReward):
 
         self.disc_lines_all_before_cascade = []
         self.n_line = None
-        self._i_am_simulate = True
+
+        # This class remembers the past state of the grid, this does not make sense for the "simulate" env
+        # so i deactivate it in this case.
+        from grid2op.Environment._ObsEnv import (
+            _ObsEnv,
+        )  # to avoid circular dependencies
+
+        self._deactivate_reward_cls = (_ObsEnv,)
 
     def initialize(self, env):
         if not env._has_attention_budget:
@@ -81,7 +88,6 @@ class _AlarmScore(AlarmReward):
         super().reset(env)
         self.window_disconnection = max(self.best_time - self.window_size, 4)
         self.disc_lines_all_before_cascade = []
-        self._i_am_simulate = self.is_simulated_env(env)
 
     def _lines_disconnected_first(self, disc_lines_at_cascading_time):
         """
@@ -111,7 +117,7 @@ class _AlarmScore(AlarmReward):
         return 1 - disc_lines_to_consider_for_score
 
     def __call__(self, action, env, has_error, is_done, is_illegal, is_ambiguous):
-        if self._i_am_simulate:
+        if self.is_simulated_env(env):
             return self.reward_no_game_over
 
         disc_lines_now = env._disc_lines
