+++ grid2op/tests/BaseBackendTest.py
@@ -57,7 +57,7 @@ from grid2op.Exceptions import *
 from grid2op.Rules import RulesChecker
 from grid2op.MakeEnv import make
 from grid2op.Rules import AlwaysLegal
-from grid2op.Action._backendAction import _BackendAction
+from grid2op.Action._BackendAction import _BackendAction
 
 import pdb
