+++ grid2op/tests/test_attached_envs.py
@@ -10,10 +10,10 @@ import warnings
 import unittest
 import grid2op
 
-from grid2op.Action.powerlineSetAction import PowerlineSetAction
+from grid2op.Action.PowerlineSetAction import PowerlineSetAction
 from grid2op.Action.PlayableAction import PlayableAction
 from grid2op.Observation.completeObservation import CompleteObservation
-from grid2op.Action.dontAct import DontAct
+from grid2op.Action.DontAct import DontAct
 from grid2op.Opponent import GeometricOpponent
 
 import pdb
